move order and cursor from paginator to specific file
