+++ cursor.go
@@ -1,0 +1,7 @@
+package paginator
+
+// Cursor cursor data
+type Cursor struct {
+	After  *string `json:"after" query:"after"`
+	Before *string `json:"before" query:"before"`
+}
