+++ order.go
@@ -1,0 +1,10 @@
+package paginator
+
+// Order type for order
+type Order string
+
+// Orders
+const (
+	ASC  Order = "ASC"
+	DESC Order = "DESC"
+)
